+++ database/migrations/2026_04_08_063612_create_mensagem_aliancas_table.php
@@ -6,17 +6,25 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
+    /**
+     * Run the migrations.
+     */
     public function up(): void
     {
-        Schema::create('mensagem_aliancas', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('alianca_id')->constrained('aliancas')->onDelete('cascade');
-            $table->foreignId('jogador_id')->constrained('jogadores')->onDelete('cascade');
-            $table->text('mensagem');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('mensagem_aliancas')) {
+            Schema::create('mensagem_aliancas', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('alianca_id')->constrained('aliancas')->onDelete('cascade');
+                $table->foreignId('jogador_id')->constrained('jogadores')->onDelete('cascade');
+                $table->text('mensagem');
+                $table->timestamps();
+            });
+        }
     }
 
+    /**
+     * Reverse the migrations.
+     */
     public function down(): void
     {
         Schema::dropIfExists('mensagem_aliancas');
